+++ packages/ovr/src/index.ts
@@ -1,7 +1,7 @@
 export { App } from "./app/index.js";
-export { Chunk } from "./chunk/index.js";
 export { Route } from "./route/index.js";
 export { Multipart } from "./multipart/index.js";
-export { jsx, render, type JSX } from "./jsx/index.js";
+export { jsx, type JSX } from "./jsx/index.js";
+export { Render } from "./render/index.js";
 export type { Middleware } from "./middleware/index.js";
 export type { Cookie } from "./cookie/index.js";